+++ src/coro_io/tests/ibverbs/CMakeLists.txt
@@ -1,7 +1,6 @@
 if(HAVE_IBVERBS_IBV)
   set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/output/tests)
   add_executable(ibverbs_test
-    test_device.cpp
     test_io.cpp
   )
   target_link_libraries(ibverbs_test -libverbs)
